Auto-reconnect if we are disconnected from device.
Add reconnect logic if we detect a disconnect.

If we are trying to run a client-submitted command, throw an exception
back to the client, who initiates the retry.

If we are passively monitoring serial for dbgprint() messages and get an
exception, the now-terminating listener thread attempts to start a new
listener thread with a fresh connection reestablished to the device.