Significantly faster ELF file loading.
Invocation of `c++filt(1)` in a subprocess takes 3--4ms, but this is
called several thousand times when parsing an ELF generated on the
arm-eabi-none architecture.

Use of this subprocess moved to a long-lived thread and names are passed
over pipes for demangling by a long-lived c++filt instance.

ARM/Thumb `empty.elf` loading time reduced from ~26s to ~5s.