CPU auto-detection now occurs if we receive a Paused msg async.
If the sketch his a breakpoint and we are notified that the sketch has
entered the debugger service while `arduino.arch == auto`, we
immediately query for the cpuid without waiting for the next
user-driven command to do so.

Switching from `auto` to a defined arduino.arch will trigger a reload of
the ELF file. (Bugfix; loading an ARM Thumb ELF without the right
ArchInterface will cause several CFI records to fail to connect to their
methods, making backtrace trigger an infinite loop.)